fix: address PR review — rename to .async convention, remove plan file
- Rename index.tsx to application-shell-splitter.async.tsx following the
  repo convention for lazy-loaded components (emmenko review)
- Update application-shell-authenticated import to use .async path
- index.tsx becomes a simple re-export
- Remove committed FEC-1017 plan file (tdeekens review)
